Clean up topics: using design 2 only
