+++ static/src/js/topic_list.js
@@ -1,5 +1,5 @@
 /***
- * Topic List Design 1
+ * Topic List Design 2
  *
  * This script visualizes a tree structure of topics and their concepts, questions and variables using the fancytree library.
  * Make sure you set study and language variable in template, else api calls will not work.
@@ -12,14 +12,17 @@
  */
 
 
-
 $.ui.fancytree.debugLevel = 0; // set debug level; 0:quiet, 1:info, 2:debug
+
+// Get current URL to read 'open' Parameter for opening specified node
+var url_string = window.location.href;
+var url = new URL(url_string);
+var open = url.searchParams.get("open");
 
 // Define buttons, which are shown when you hover over a topic or concept
 // This buttons will be append to the nodes defined by fancytree
 var filter_options_string = "<span class='btn-group btn-group-sm filter-options' role='group'><button type='button' data-tooltip='tooltip' title='Show all related variables' onclick='filter(this, \"variable\")' class='btn btn-link filter-option-variable' ><span class='glyphicon glyphicon-stats' aria-hidden='true'></span></button><button type='button' class='btn btn-link filter-option-question' data-tooltip='tooltip' title='Show all related questions' onclick='filter(this, \"question\")'><span class='glyphicon glyphicon-question-sign' aria-hidden='true'></span></button>" +
     "<button type='button' data-tooltip='tooltip' title='Add all related variables to one of your baskets' onclick='addToBasket(this)' class='btn btn-link' data-toggle='modal' data-target='#topic-list-add-to-basket'><span class='glyphicon glyphicon-shopping-cart' aria-hidden='true'></span></button></span>"
-var filter_options_var_quest_string = "<span class='btn-group btn-group-sm filter-options' role='group'><button type='button' data-tooltip='tooltip' title='Go to detail view' class='btn btn-link' onclick='redirect(this)'><span class='glyphicon glyphicon-circle-arrow-right' aria-hidden='true'></span></button><button type='button' data-tooltip='tooltip' title='Add variables to one of your baskets' onclick='addToBasket(this)' class='btn btn-link btn-sm' data-toggle='modal' data-target='#topic-list-add-to-basket'><span class='glyphicon glyphicon-shopping-cart' aria-hidden='true'></span></button></span>"
 var base_url = location.protocol + '//' + window.location.host + "/api/topics/" + study + "/" + language
 
 // Define what the tree structure will look like, for more information and options see https://github.com/mar10/fancytree.
@@ -33,7 +36,7 @@ $(function () {
             "variable": {icon: "glyphicon glyphicon-stats"},
             "question": {icon: "glyphicon glyphicon-question-sign"},
         },
-        filter: {  // override default settings
+        filter: {
             counter: false, // No counter badges
             mode: "hide"  // "dimm": Grayout unmatched nodes, "hide": remove unmatched nodes
         },
@@ -41,8 +44,6 @@ $(function () {
             return data.typeInfo.icon;
         },
         glyph: {
-            // The preset defines defaults for all supported icon types.
-            // It also defines a common class name that is prepended (in this case 'fa ')
             preset: "bootstrap3",
             map: {
                 _addClass: "glyphicon",
@@ -70,7 +71,7 @@ $(function () {
 
         },
         source: {
-            url: base_url,
+            url: base_url, // load data from api (topic and concepts only)
             cache: false
         },
         renderNode: function (event, data) {
@@ -80,21 +81,19 @@ $(function () {
             if ($(node.span).find('span.filter-options').length == 0) {
                 if (node.type == 'topic' || node.type == 'concept') {
                     $spanTitle.after(filter_options_string);
-                }
-                else if (node.type == 'variable' || node.type == 'question') {
-                    $spanTitle.after(filter_options_var_quest_string);
+                    // $spanTitle.before(node.key)
                 }
             }
-
         },
         // When tree fully loaded: if parameter 'open' is set in URL open specified node
-        init: function(event, data){
-            if(open != null) {
+        init: function (event, data) {
+            if (open != null) {
                 var node = $("#tree").fancytree("getNodeByKey", open);
                 node.makeVisible();
             }
         }
     });
+
 
     // Search the tree for search string
     $('#btn-search').on('click', function () {
@@ -102,108 +101,46 @@ $(function () {
     });
 
     // Trigger search on enter
-    $('.search-bar').keypress(function(e){
-        if(e.which == 13){//Enter key pressed
+    $('.search-bar').keypress(function (e) {
+        if (e.which == 13) {//Enter key pressed
             $('#btn-search').click();
         }
     });
 
-    // Activate tooltip for more information about filter buttons, e.g. 'Show all related variables'
-    $('body').tooltip({selector:'[data-tooltip=tooltip]', trigger: "hover"});
 
+    // Activate tooltip for more information about filter buttons, e.g. 'Show all related variables'
+    $('body').tooltip({selector: '[data-tooltip=tooltip]', trigger: "hover"});
 });
 
+
 // On click on a topic or concept show all variables oder questions
-// Pay attention to the different states of the tree (has children / has no children / children already shown)
 function filter(node, type) {
+
+    // show spinner while loading
+    $("#tree_variables").empty();
+    $('.sk-three-bounce').show();
+    $('#tree').find("button[class*='-btn-active']").removeClass("variable-btn-active question-btn-active")
+    $(node).toggleClass(type + '-btn-active')
 
     var activeNode = $.ui.fancytree.getNode(node);
 
     var extraClasses = activeNode.extraClasses || "";
-    var url = base_url + "/" + activeNode.key;
-    if(type == 'variable'){
+    var url = base_url + "/" + activeNode.key
+    if (type == 'variable') {
         url += '?variable_html=true';
-    } if(type == 'question'){
+    }
+    if (type == 'question') {
         url += '?question_html=true';
     }
     var data;
-    jQuery.getJSON(url, function (json) {
-        data = json;
-        if (activeNode.hasChildren() && !$(activeNode.li).hasClass(type + '-btn-active')) {
-            removeAllChildren(activeNode, type);
-            if (type == 'variable') {
-                for (var i = 0; i < data.variable_list.length; i++) {
-                    var concept_key = data.variable_list[i].concept_key.toString();
-                    data.variable_list[i]['extraClasses'] = 'async-data-' + type;
-                    var parentNode = $("#tree").fancytree("getNodeByKey", concept_key);
-                    parentNode.addChildren(data.variable_list[i]);
-                    // $(parentNode).addClass(type + '-btn-active')
-                }
-            }
-            if (type == 'question') {
-                if (data.question_list) {
-                    for (var i = 0; i < data.question_list.length; i++) {
-                        var concept_key = data.question_list[i].concept_key.toString();
-                        data.question_list[i]['extraClasses'] = 'async-data-' + type;
-                        var parentNode = $("#tree").fancytree("getNodeByKey", concept_key);
-                        parentNode.addChildren(data.question_list[i]);
-                    }
-                }
-            }
-            $(activeNode.li).addClass(type + '-btn-active')
-            activeNode.visit(function (node) {
-                node.removeClass('fancytree-hide')
-                node.setExpanded(true);
-                $(node.li).addClass(type + '-btn-active')
-            }, true);
 
-            activeNode.setExpanded(true);
-        }
-        else if (activeNode.hasChildren() && $(activeNode.li).hasClass(type + '-btn-active')) {
-            removeAllChildren(activeNode, type);
-            $(activeNode.li).removeClass(type + '-btn-active')
-            activeNode.visit(function (node) {
-                if (node.type == 'topic' || node.type == 'concept') {
-                    $(node.li).removeClass(type + '-btn-active')
-                }
-            }, true);
-        }
-        else if (!activeNode.hasChildren() && !$(activeNode.li).hasClass(type + '-btn-active')) {
-            if (type == 'variable') {
-                for (var i = 0; i < data.variable_list.length; i++) {
-                    var concept_key = data.variable_list[i].concept_key.toString();
-                    data.variable_list[i]['extraClasses'] = 'async-data-' + type;
-                    var parentNode = $("#tree").fancytree("getNodeByKey", concept_key);
-                    parentNode.addChildren(data.variable_list[i]);
-                    $(parentNode.li).addClass(type + '-btn-active')
-                    parentNode.setExpanded(true)
-                }
-            }
-            if (type == 'question') {
-                if (data.question_list) {
-                    for (var i = 0; i < data.question_list.length; i++) {
-                        var concept_key = data.question_list[i].concept_key.toString();
-                        data.question_list[i]['extraClasses'] = 'async-data-' + type;
-                        var parentNode = $("#tree").fancytree("getNodeByKey", concept_key);
-                        parentNode.addChildren(data.question_list[i]);
-                        parentNode.setExpanded(true)
-                    }
-                }
-            }
-            $(activeNode).toggleClass(type + '-btn-active')
-            activeNode.visit(function (node) {
-                node.removeClass('fancytree-hide')
-                node.setExpanded(true);
-                $(node.li).addClass(type + '-btn-active')
-            }, true);
-
-            activeNode.setExpanded(true);
-        }
-        else if (!activeNode.hasChildren() && $(activeNode).hasClass(type + '-btn-active')) {
-            removeAllChildren(activeNode, type);
-            $(activeNode).removeClass(type + '-btn-active')
-        }
-
+    jQuery.get(url, function (data) {
+        $('.sk-three-bounce').hide(); // hide the loading message
+        $("#tree_variables").html(data);
+        $("#variable_table").DataTable();
+    }).fail(function () {
+        $('.sk-three-bounce').hide(); // hide the loading message
+        $("#tree_variables").html("<p><span class='glyphicon glyphicon-alert' aria-hidden='true'></span> Load Error!</p>");
     });
 }
 
@@ -271,8 +208,8 @@ function addToBasket(el) {
 
     var url = base_url + "/baskets";
     jQuery.getJSON(url, function (data) {
-        if(data.user_logged_in) {
-            if(data.baskets.length == 0){
+        if (data.user_logged_in) {
+            if (data.baskets.length == 0) {
                 var redirect_create_basket_url = location.protocol + '//' + window.location.host + "/workspace/baskets";
                 $('#basket_list').append("<p><a class='btn btn-primary' href='" + redirect_create_basket_url + "'>Create a basket</a></p>");
             }
@@ -304,13 +241,3 @@ $('#topic-list-add-to-basket').on('hidden.bs.modal', function () {
     $('#basket_success').addClass('hidden');
     $('#basket_error').addClass('hidden');
 })
-
-
-// redirect to detail page of element
-function redirect(el) {
-    var node = $.ui.fancytree.getNode(el);
-    var url = location.protocol + '//' + window.location.host + "/api/topics/redirect/" + node.key;
-    window.location.href = url;
-}
-
-
