+++ templates/studies/study_topics.html
@@ -9,8 +9,8 @@
 <!-- TODO: Load jquery-fancytree from webpack here -->
 
 <script>
-    var study = '{{study.name}}';
-    var language = '{{language}}';
+    var study = '{{study.name}}'
+    var language = '{{ language }}'
 </script>
 <script src="./../../static/src/js/topic_list.js"></script>
 {% endblock scripts %}
@@ -18,24 +18,33 @@
 {% block content %}
 
 <h2><span class="small glyphicon glyphicon-folder-close"></span> Topics</h2>
-<div class="conainer test">
-    <a class="anchor" id="topics"></a>
-    <div class="row">
-        <div class="col-md-offset-8 col-md-4 col-xs-8 col-xs-offset-4 search-bar">
-            <div class="input-group">
-                <input type="text" class="form-control" placeholder="Search node .." id="search">
-                <span class="input-group-btn">
+<a class="anchor" id="topics"></a>
+<div class="row">
+    <div class="col-md-offset-8 col-md-4 col-xs-8 col-xs-offset-4 search-bar">
+        <div class="input-group">
+            <input type="text" class="form-control" placeholder="Search node .." id="search">
+            <span class="input-group-btn">
 		<button class="btn btn-default" type="button" aria-pressed="true" id="btn-search"><span
                 class='glyphicon glyphicon-search'
                 aria-hidden='true'></span></button>
 	</span>
-            </div>
         </div>
     </div>
-    <div class="row topic-list-wrapper">
-        <div class="col-md-12 col-xs-12">
-            <div class="topic-list-content">
-                <div id="tree"></div>
+</div>
+<div class="row topic-list-wrapper">
+    <div class="col-md-6 col-xs-6">
+        <div class="topic-list-content">
+            <div id="tree"></div>
+        </div>
+    </div>
+    <div class="col-md-6 col-xs-6">
+        <div class="topic-list-content">
+            <div class="sk-three-bounce" style="display: none">
+                <div class="sk-child sk-bounce1"></div>
+                <div class="sk-child sk-bounce2"></div>
+                <div class="sk-child sk-bounce3"></div>
+            </div>
+            <div id="tree_variables">
             </div>
         </div>
     </div>
